pyproject.toml
[project]
name = "AreUSleepy"
version = "0.1.0"
description = "基于 Sleepy 状态查询的插件"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-plugin-apscheduler>=0.3.0",
    "nonebot-plugin-alconna>=0.12.0",
    "nonebot2[httpx]>=2.4.2",
    "nonebot2[fastapi]>=2.4.2",
    "nonebot2[websockets]>=2.4.2",
    "nonebot_adapter_onebot>=2.4.6"
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugin_dirs = ["./plugins"]
builtin_plugins = []

